Union-resolve native_backend (keep main's LAS parity fallback alongside
the origin classification) and delete both committed root cpython-313
binaries per #435. Also drop the bare geo-viz-engine root from the
pytest pythonpath: it hosts the submodule's own committed .so binaries
which shadow freshly built extensions the same way the repo-root
binaries did (observed live: the map-render 3.13 leg failed audit
parity through exactly this path). geoviz itself resolves from the
editable installs in requirements-geoviz.txt.

+++ native/grid_render_core/src/bindings.cpp
@@ -60,7 +60,7 @@ std::vector<std::uint8_t> ramp_values(const U8Array& ramp) {
 }
 
 py::array_t<std::uint8_t> render_grid_rgba_py(
-    FloatArray grid_z, py::object mask_obj, U8Array lut, float lo, float hi,
+    FloatArray grid_z, py::object mask_obj, U8Array lut, double lo, double hi,
     float gamma, std::uint8_t opacity) {
     auto gz = grid_z.request();
     if (gz.ndim != 2) {
